Create component QuizButton that navigates to quizes
